+++ src/Resource/Limits.php
@@ -6,6 +6,9 @@ namespace DigitalCz\DigiSign\Resource;
 
 class Limits extends BaseResource
 {
+    public ?int $envelopes;
+
+    /** @deprecated Use $envelopes instead */
     public ?int $envelopesMonthly;
 
     public ?int $users;
@@ -16,11 +19,19 @@ class Limits extends BaseResource
 
     public ?int $documents;
 
+    public ?int $noneOrManualIdentifications;
+
+    /** @deprecated Use $noneOrManualIdentifications instead */
     public ?int $noneOrManualMonthlyIdentifications;
 
+    public ?int $aiIdentifications;
+
+    /** @deprecated Use $aiIdentifications instead */
     public ?int $aiMonthlyIdentifications;
 
     public ?int $tags;
 
     public ?int $fileSize;
+
+    public string $frequency;
 }